requirements.txt
requests>=2.28.1
selenium>=4.4.0
beautifulsoup4>=4.11.1
pymongo>=4.2.0
pytz>=2022.2.1
python-dateutil>=2.8.2
newspaper3k>=0.2.8
webdriver-manager>=3.8.3
